Incorporated Rich module for better output amoung other improvements

+++ dev_main.py
@@ -10,17 +10,18 @@ from subprocess import Popen, PIPE
 from datetime import datetime
 from typing import List, Dict, Tuple, Optional
 import threading
-from termcolor import colored
-from utils import bad_print, bold_print, selection_print, is_member_of_group
+from utils import print_banner, rich_important_print, rich_bad_print, rich_selection_print, rich_bold_print, rich_print 
+from utils import rich_success_print, is_member_of_group
 from utils import add_to_log, read_pid_lock, write_pid_lock, get_time_lapsed, get_customer_prefix_list
 from utils import send_to_devices, select_action, get_prefixes, generate_commands, push_changes
 
 
+# Author: Richard Blackwell
+# Date: 1 August 2024 
+# Version: 0.2.0
 
-#Author: Richard Blackwell
-#Date: 1 August 2024 
-#Version: 0.1.0
-
+# 08/1/2024 - 0.1.0 - Initial version of the script
+# 08/31/2024 - 0.2.0 - Incorporated the Rich module for all output
 
 
 test_mode = False # test_mode will ensure that the script only runs on the test devices
@@ -43,41 +44,31 @@ alu_cmds_file_path = ('./__cmds_file_alu__.log')
 jnpr_cmds_file_path = ('./__cmds_file_jnpr__.log')
 
 
-lumen_banner = """
-****************************************************************************************
-            
-                     ██╗     ██╗   ██╗███╗   ███╗███████╗███╗   ██╗                  
-                     ██║     ██║   ██║████╗ ████║██╔════╝████╗  ██║
-                     ██║     ██║   ██║██╔████╔██║█████╗  ██╔██╗ ██║
-                     ██║     ██║   ██║██║╚██╔╝██║██╔══╝  ██║╚██╗██║
-                     ███████╗╚██████╔╝██║ ╚═╝ ██║███████╗██║ ╚████║
-                     ╚══════╝ ╚═════╝ ╚═╝     ╚═╝╚══════╝╚═╝  ╚═══╝        
-                                                                                   
-                  ****************************************************               
-                  *                      Lumen                       *               
-                  *                     Security                     *               
-                  *                                                  *               
-                  *        DDoS AORC Modify Prefix-List Script       *               
-                  *                                                  *               
-                  *    For issues with this script, please reach     *               
-                  *             out to Richard Blackwell             *               
-                  *                                                  *               
-                  *           Richard.Blackwell@lumen.com            *               
-                  ****************************************************                  
-                                                                                    
-****************************************************************************************
-"""
+lumen_banner = f"""
+██╗     ██╗   ██╗███╗   ███╗███████╗███╗   ██╗
+██║     ██║   ██║████╗ ████║██╔════╝████╗  ██║
+██║     ██║   ██║██╔████╔██║█████╗  ██╔██╗ ██║
+██║     ██║   ██║██║╚██╔╝██║██╔══╝  ██║╚██╗██║
+███████╗╚██████╔╝██║ ╚═╝ ██║███████╗██║ ╚████║
+╚══════╝ ╚═════╝ ╚═╝     ╚═╝╚══════╝╚═╝  ╚═══╝
 
-horiz_line = "----------------------------------------------------------------------------------------"
+╭──────────────────────────────────────────────────╮
+│               Lumen DDoS Security                │
+│                                                  │
+│        DDoS AORC Modify Prefix-List Script       │
+│                                                  │
+│    For issues with this script, please reach     │
+│             out to Richard Blackwell             │
+│                                                  │
+│           Richard.Blackwell@lumen.com            │
+╰──────────────────────────────────────────────────╯"""
 
-script_banner = (
-    f"{horiz_line}\n"
-    "                   The purpose of this script is to allow the DDoS SOC                  \n"
-    "           to add and remove prefixes from existing AORC customer's policies            \n\n"
-    "                           additional info can be found here:                           \n"
-    "https://nsmomavp045b.corp.intranet:8443/display/SOPP/AORC+Only+Modify+Prefix-list+Script\n"       
-    f"{horiz_line}\n"
-)
+script_banner = f"""
+The purpose of this script is to allow the DDoS SOC
+to add and remove prefixes from existing AORC customer's policies
+
+Additional info can be found here:
+https://nsmomavp045b.corp.intranet:8443/display/SOPP/AORC+Only+Modify+Prefix-list+Script"""
 
 
 # One Nokia and one Juniper spare device that ROCI works with. These devices do not particpate with the local scrubbers.
@@ -89,24 +80,24 @@ test_devices = [
 
 # Complplete list of production PE routers that participate with the local scrubbers. 
 prod_devices = [
-    {'manufacturer': 'Juniper', 'dns': 'edge9.sjo1'},
-    {'manufacturer': 'Juniper', 'dns': 'edge3.chi10'},
-    {'manufacturer': 'Juniper', 'dns': 'edge3.syd1'},
-    {'manufacturer': 'Nokia',   'dns': 'ear3.ams1'},
-    {'manufacturer': 'Nokia',   'dns': 'msr2.frf1'},
-    {'manufacturer': 'Nokia',   'dns': 'msr3.frf1'},
-    {'manufacturer': 'Nokia',   'dns': 'msr11.hkg3'},
-    {'manufacturer': 'Nokia',   'dns': 'msr12.hkg3'},
-    {'manufacturer': 'Nokia',   'dns': 'msr2.lax1'},
-    {'manufacturer': 'Nokia',   'dns': 'msr3.lax1'},
-    {'manufacturer': 'Nokia',   'dns': 'ear4.lon2'},
-    {'manufacturer': 'Nokia',   'dns': 'msr1.nyc1'},
-    {'manufacturer': 'Nokia',   'dns': 'ear2.par1'},
-    {'manufacturer': 'Nokia',   'dns': 'msr11.sap1'},
-    {'manufacturer': 'Nokia',   'dns': 'msr12.sap1'},
-    {'manufacturer': 'Nokia',   'dns': 'msr11.sng3'},
-    {'manufacturer': 'Nokia',   'dns': 'msr12.sng3'},
-    {'manufacturer': 'Nokia',   'dns': 'msr11.tok4'},
+    # {'manufacturer': 'Juniper', 'dns': 'edge9.sjo1'},
+    # {'manufacturer': 'Juniper', 'dns': 'edge3.chi10'},
+    # {'manufacturer': 'Juniper', 'dns': 'edge4.syd1'},
+    # {'manufacturer': 'Nokia',   'dns': 'ear3.ams1'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr2.frf1'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr3.frf1'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr11.hkg3'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr12.hkg3'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr2.lax1'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr3.lax1'},
+    # {'manufacturer': 'Nokia',   'dns': 'ear4.lon2'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr1.nyc1'},
+    # {'manufacturer': 'Nokia',   'dns': 'ear2.par1'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr11.sap1'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr12.sap1'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr11.sng3'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr12.sng3'},
+    # {'manufacturer': 'Nokia',   'dns': 'msr11.tok4'},
     {'manufacturer': 'Nokia',   'dns': 'msr2.wdc12'},
     {'manufacturer': 'Nokia',   'dns': 'msr3.wdc12'},
     {'manufacturer': 'Nokia',   'dns': 'msr1.dal1'}
@@ -120,14 +111,12 @@ def cleanup_files() -> None:
         if os.path.exists(jnpr_cmds_file_path):
             os.remove(jnpr_cmds_file_path)
     except OSError as e:
-        print(f"Error deleting files: {e}")
+        rich_bad_print(f"Error deleting files: {e}")
 
 
 # Function to Kill the program after the time limit has been reached
 def timeout():
-    print(f"\n{horiz_line}")
-    bad_print(f"Time's up! This program has a time limit of {total_time_limit/60:.2f} minutes.")
-    print("Exiting program...")
+    rich_bad_print(f"Time's up! This program has a time limit of {int(total_time_limit/60)} minute(s). Exiting program.")
     os._exit(1)
 
 
@@ -161,22 +150,22 @@ def lock_resource():
                     
                     time_lapsed = get_time_lapsed(info["Timestamp"])
                     # Print the message to the user including the user who is running the script and the time lapsed
-                    print(f"\n{horiz_line}")
-                    bad_print("This program is already in use. Only one instance of this script can be run at one time.")
-                    bad_print(f"User '{info['Username']}' is already running this program. Time lapsed: '{str(time_lapsed)[:8]}'\n{horiz_line}")
-                    print("")
+                    already_runner_banner = (
+                            f"\nThis program is already in use. Only one instance of this script can be run at one time.\n"
+                            f"User '{info['Username']}' is already running this program. Time lapsed: {str(time_lapsed)[:8]}\n")
+                    rich_bad_print(already_runner_banner)
 
                     # If the user wants to retry, wait for the specified time before retrying
                     if attempt < attempt_limit - 1:
                         retry = input("Do you want to try again? (Y/YES to retry): ").strip().upper()
                         if retry not in ['Y', 'YES']:
-                            print("Exiting program...")
+                            rich_bad_print("Exiting program...")
                             sys.exit(1)
                         print(f"Retrying in {wait_time} seconds...")
                         time.sleep(wait_time)
                     else:
-                        print(f"Failed to acquire lock after {attempt_limit} attempts.")
-                        print("You have reached the maximum number of attempts. Exiting.")
+                        rich_bad_print(f"Failed to acquire lock after {attempt_limit} attempts.\n"
+                                       "You have reached the maximum number of attempts. Exiting program.")
                         sys.exit(1)
                 
                 # Release the lock
@@ -200,7 +189,8 @@ def main() -> None:
         # Check if the user is a member of the group "ddos_ops"
         username = os.getlogin()
         if not is_member_of_group(group_name):
-            print(f"You do not have sufficient permission to run this program. User '{username}' must be a member of the '{group_name}' group.")
+            rich_bad_print(f"You do not have sufficient permission to run this program.\n"
+                           "User '{username}' must be a member of the '{group_name}' group.")
             sys.exit(1)
 
         # Set the devices to be used based on the test_mode flag
@@ -217,18 +207,19 @@ def main() -> None:
         user_decisions = add_to_log(log_file_path, user_decisions, "Timestamp", datetime.now().strftime('%Y-%m-%d %H:%M:%S'))
 
         # Print the banner
-        print(lumen_banner)
-        if test_mode: bad_print("TEST MODE: Reminder that you are in test mode.")
-        if dryrun: bad_print("DRYRUN MODE: Reminder that you are in dryrun mode.")   
-        bold_print(f"\n{script_banner}")
+        print_banner(lumen_banner)
+        print()
+        if test_mode: rich_important_print("TEST MODE: Reminder that you are in test mode.")
+        if dryrun: rich_important_print("DRYRUN MODE: Reminder that you are in dryrun mode.")   
+        print_banner(script_banner)
 
         # Get the customer's prefix-list
         selected_policy: str
         cust_id: str
         selected_policy, cust_id = get_customer_prefix_list(prod_devices, alu_cmds_file_path, jnpr_cmds_file_path, dryrun)
-        selection_print(f"\nYou have selected: {selected_policy}")
         user_decisions = add_to_log(log_file_path, user_decisions, "Provided Cust ID", cust_id)
         user_decisions = add_to_log(log_file_path, user_decisions, "Selected policy", selected_policy)
+        rich_selection_print(f"You have selected: {selected_policy}")
 
         # Add or remove prefixes
         menu: Dict[str, str] = {  
@@ -236,21 +227,24 @@ def main() -> None:
             "2": "Remove prefixes"}
         action: str = select_action(menu)
         if action == "1":
-            selection_print("\nYou have selected: Add prefixes")
+            rich_selection_print("You have selected: Add prefixes")
             user_decisions = add_to_log(log_file_path, user_decisions, "Selected action", "Add prefixes")
             add_prefixes: bool = True
             remove_prefixes: bool = False
         elif action == "2":
-            selection_print("\nYou have selected: Remove prefixes")
-            user_decisions = add_to_log(log_file_path, user_decisions, "Selected action", "Add prefixes")
+            rich_selection_print("You have selected: Remove prefixes")
+            user_decisions = add_to_log(log_file_path, user_decisions, "Selected action", "Remove prefixes")
             add_prefixes: bool = False
             remove_prefixes: bool = True
+        else:
+            rich_bad_print(f"Invalid action selection. '{action}' was selected. Please report this error. Exiting program...")
+            sys.exit(1)
 
         # Get the prefixes from the user
         valid_prefixes: List[str]
         prefix_confirm: str
         valid_prefixes, prefix_confirm = get_prefixes()
-        selection_print("\nUser has confirmed the prefixes. Proceeding with generating commands...")
+        rich_selection_print("User has confirmed the prefixes. Proceeding with generating commands...")
         user_decisions = add_to_log(log_file_path, user_decisions, "Provided Prefixes", valid_prefixes)
         user_decisions = add_to_log(log_file_path, user_decisions, "Prefix confirmation", prefix_confirm)
         
@@ -260,10 +254,13 @@ def main() -> None:
         config_confirm: str
         cmds_alu, cmds_jnpr, config_confirm = generate_commands(valid_prefixes, selected_policy, add_prefixes, remove_prefixes, test_mode, dryrun)
 
-        selection_print(f"\nUser has confirmed the commands. Proceeding with pushing the commands to the devices...")
         user_decisions = add_to_log(log_file_path, user_decisions, "Nokia Configuration", cmds_alu)
         user_decisions = add_to_log(log_file_path, user_decisions, "Juniper Configuration", cmds_jnpr)
         user_decisions = add_to_log(log_file_path, user_decisions, "Configuration confirmation", config_confirm)
+        if not config_confirm: 
+            rich_selection_print("User has chosen not confimed the configuration changes and the changes will not be pushed to producation devices. Exiting program...")
+            sys.exit(0)
+        rich_selection_print("User has confirmed the configuration. Proceeding with pushing the changes to the devices...")
 
         # Write the commands to the configuration files
         with open(alu_cmds_file_path, 'w+') as file:
@@ -276,12 +273,11 @@ def main() -> None:
             file.close()
         if config_confirm: 
                 output = send_to_devices(push_changes, devices, alu_cmds_file_path, jnpr_cmds_file_path, dryrun)
-        print(f"\n{horiz_line}")
-        selection_print("Configuration changes have been pushed to the devices successfully!")
-        print(f"{horiz_line}\n")
+        print("")
+        rich_success_print(f"\nConfiguration changes have been pushed to the devices successfully!\n")
 
         # Print the log of user's decisions
-        selection_print(f"Log of user's decisions:\n")
+        rich_selection_print(f"Decision log:")
         for key, value in user_decisions.items():
             if key == "Nokia Configuration" or key == "Juniper Configuration":
                 print(f"{key}:")
@@ -291,9 +287,7 @@ def main() -> None:
             else: 
                 print(f"{key}: {value}")
 
-        
-        print(f"\n{horiz_line}")
-        selection_print("Program Finished. Exiting...")
+        rich_selection_print("Program Finished. Exiting...")
         sys.exit(0)
     except KeyboardInterrupt:
         print("\nProcess interrupted by user. Exiting program...")
